ChatBot_PS-1/database_chatbot/requirements.txt
alabaster==0.7.12
asgiref==3.2.7
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
dataclasses==0.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.7
django-import-export==2.2.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
filelock==3.0.12
future==0.18.2
idna==2.9
imagesize==1.2.0
jdcal==1.4.1
joblib==0.15.1
MarkupPy==1.14
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysqlclient==1.4.6
nltk==3.5
numpy==1.18.5
odfpy==1.4.1
openpyxl==3.0.3
packaging==20.4
Pygments==2.6.1
PyMySQL==0.9.3
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.3.1
regex==2020.6.8
requests==2.23.0
sacremoses==0.0.43
scikit-learn==0.23.1
scipy==1.4.1
sentence-transformers==0.2.6.1
sentencepiece==0.1.91
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
threadpoolctl==2.1.0
tokenizers==0.7.0
torch==1.5.0
tqdm==4.46.1
transformers==2.11.0
urllib3==1.25.9
xlrd==1.2.0
xlwt==1.3.0